fp get reception api
